fix legacy ntp
